+++ packages/expo-cli/src/commands/eject/Eject.ts
@@ -1,5 +1,5 @@
 import { ExpoConfig, getConfig, PackageJSONConfig } from '@expo/config';
-import { WarningAggregator } from '@expo/config-plugins';
+import { ModPlatform, WarningAggregator } from '@expo/config-plugins';
 import { getBareExtensions, getFileWithExtensions } from '@expo/config/paths';
 import JsonFile, { JSONObject } from '@expo/json-file';
 import { Exp } from '@expo/xdl';
@@ -26,13 +26,13 @@ import maybeBailOnGitStatusAsync from '../utils/maybeBailOnGitStatusAsync';
 import { getOrPromptForBundleIdentifier, getOrPromptForPackage } from './ConfigValidation';
 
 type DependenciesMap = { [key: string]: string | number };
-type PlatformsArray = ('ios' | 'android')[];
 
 export type EjectAsyncOptions = {
   verbose?: boolean;
   force?: boolean;
   install?: boolean;
   packageManager?: 'npm' | 'yarn';
+  platforms: ModPlatform[];
 };
 
 /**
@@ -47,21 +47,15 @@ export type EjectAsyncOptions = {
  */
 export async function ejectAsync(
   projectRoot: string,
-  options: EjectAsyncOptions = {}
+  { platforms, ...options }: EjectAsyncOptions
 ): Promise<void> {
-  if (await maybeBailOnGitStatusAsync()) return;
-
-  const platforms: PlatformsArray = ['android'];
-
-  // Skip ejecting for iOS on Windows
-  if (process.platform !== 'win32') {
-    platforms.push('ios');
+  if (!platforms?.length) {
+    throw new CommandError('At least one platform must be enabled when syncing');
   }
 
-  const { exp, pkg } = await ensureConfigAsync(projectRoot);
-  const tempDir = temporary.directory();
-
-  if (!platforms.includes('ios')) {
+  const isWindows = process.platform === 'win32';
+  // Skip ejecting for iOS on Windows
+  if (isWindows && !platforms.includes('ios')) {
     log.warn(
       `⚠️  Skipping generating the iOS native project files. Run ${chalk.bold(
         'expo eject'
@@ -69,6 +63,11 @@ export async function ejectAsync(
     );
     log.newLine();
   }
+
+  if (await maybeBailOnGitStatusAsync()) return;
+
+  const { exp, pkg } = await ensureConfigAsync({ projectRoot, platforms });
+  const tempDir = temporary.directory();
 
   const { hasNewProjectFiles, needsPodInstall } = await createNativeProjectsFromTemplateAsync({
     projectRoot,
@@ -96,11 +95,11 @@ export async function ejectAsync(
 
   // Apply Expo config to native projects
   if (platforms.includes('ios')) {
-    await configureIOSStepAsync(projectRoot);
+    await configureIOSStepAsync({ projectRoot, platforms });
   }
 
   if (platforms.includes('android')) {
-    await configureAndroidStepAsync(projectRoot);
+    await configureAndroidStepAsync({ projectRoot, platforms });
   }
 
   // Install CocoaPods
@@ -112,7 +111,7 @@ export async function ejectAsync(
     log.debug('Skipped pod install');
   }
 
-  await warnIfDependenciesRequireAdditionalSetupAsync(pkg, options);
+  await warnIfDependenciesRequireAdditionalSetupAsync(pkg);
 
   log.newLine();
   log.nested(`➡️  ${chalk.bold('Next steps')}`);
@@ -191,9 +190,9 @@ export async function ejectAsync(
   }
 }
 
-async function configureIOSStepAsync(projectRoot: string) {
+async function configureIOSStepAsync(props: { projectRoot: string; platforms: ModPlatform[] }) {
   const applyingIOSConfigStep = CreateApp.logNewSection('iOS config syncing');
-  await configureIOSProjectAsync(projectRoot);
+  await configureIOSProjectAsync(props);
   if (WarningAggregator.hasWarningsIOS()) {
     applyingIOSConfigStep.stopAndPersist({
       symbol: '⚠️ ',
@@ -240,9 +239,9 @@ async function installNodeDependenciesAsync(
   }
 }
 
-async function configureAndroidStepAsync(projectRoot: string) {
+async function configureAndroidStepAsync(props: { projectRoot: string; platforms: ModPlatform[] }) {
   const applyingAndroidConfigStep = CreateApp.logNewSection('Android config syncing');
-  await configureAndroidProjectAsync(projectRoot);
+  await configureAndroidProjectAsync(props);
   if (WarningAggregator.hasWarningsAndroid()) {
     applyingAndroidConfigStep.stopAndPersist({
       symbol: '⚠️ ',
@@ -273,9 +272,13 @@ function copyPathsFromTemplate(
   return [copiedPaths, skippedPaths];
 }
 
-async function ensureConfigAsync(
-  projectRoot: string
-): Promise<{ exp: ExpoConfig; pkg: PackageJSONConfig }> {
+async function ensureConfigAsync({
+  projectRoot,
+  platforms,
+}: {
+  projectRoot: string;
+  platforms: ModPlatform[];
+}): Promise<{ exp: ExpoConfig; pkg: PackageJSONConfig }> {
   // We need the SDK version to proceed
 
   let exp: ExpoConfig;
@@ -305,15 +308,20 @@ async function ensureConfigAsync(
 
   // Prompt for the Android package first because it's more strict than the bundle identifier
   // this means you'll have a better chance at matching the bundle identifier with the package name.
-  await getOrPromptForPackage(projectRoot);
-  await getOrPromptForBundleIdentifier(projectRoot);
+  if (platforms.includes('android')) {
+    await getOrPromptForPackage(projectRoot);
+  }
+  if (platforms.includes('ios')) {
+    await getOrPromptForBundleIdentifier(projectRoot);
+  }
 
   if (exp.entryPoint) {
     delete exp.entryPoint;
     log(`- expo.entryPoint is not needed and has been removed.`);
   }
 
-  return { exp, pkg };
+  // Read config again because prompting for bundle id or package name may have mutated the results.
+  return getConfig(projectRoot);
 }
 
 function createFileHash(contents: string): string {
@@ -543,7 +551,7 @@ export function hashForDependencyMap(deps: DependenciesMap): string {
 export function getTargetPaths(
   projectRoot: string,
   pkg: PackageJSONConfig,
-  platforms: PlatformsArray
+  platforms: ModPlatform[]
 ) {
   const targetPaths: string[] = [...platforms];
 
@@ -573,7 +581,7 @@ async function cloneNativeDirectoriesAsync({
   tempDir: string;
   exp: Pick<ExpoConfig, 'name' | 'sdkVersion'>;
   pkg: PackageJSONConfig;
-  platforms: PlatformsArray;
+  platforms: ModPlatform[];
 }): Promise<string[]> {
   const templateSpec = await validateBareTemplateExistsAsync(exp.sdkVersion!);
 
@@ -642,7 +650,7 @@ async function createNativeProjectsFromTemplateAsync({
   exp: ExpoConfig;
   pkg: PackageJSONConfig;
   tempDir: string;
-  platforms: PlatformsArray;
+  platforms: ModPlatform[];
 }): Promise<
   { hasNewProjectFiles: boolean; needsPodInstall: boolean } & DependenciesModificationResults
 > {
@@ -702,8 +710,7 @@ function createDependenciesMap(dependencies: any): DependenciesMap {
  * users to add some code, eg: to their AppDelegate.
  */
 async function warnIfDependenciesRequireAdditionalSetupAsync(
-  pkg: PackageJSONConfig,
-  options?: EjectAsyncOptions
+  pkg: PackageJSONConfig
 ): Promise<void> {
   const expoPackagesWithExtraSetup = [
     'expo-camera',